Optimiser in editor and interface changes
I added the optimised tree to the editor debug output.
I also started on the refactoring to make the instruction layout more
decoupled from the CPU and assembler. The instruction layout should be
basically opaque to the assembler and the cpu. That way we can easily
support more complex instruction encodings.